# opendma-api-spec-070.yaml
openapi: 3.0.4
info:
  title: OpenDMA
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.7.0
tags:
  - name: Service
  - name: Objects
  - name: Content
paths:
  /opendma:
    get:
      tags:
        - Service
      summary: Information about this OpenDMA service
      description: Get information about this OpenDMA service endpoint.
      operationId: getServiceDescriptor
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDescriptor'
  /opendma/obj/{repoId}:
    get:
      tags:
        - Objects
      summary: Repository Info
      description: Returns the Repository object describing this repository.
      operationId: getRepositoryInfo
      parameters:
        - name: repoId
          in: path
          description: ID of Repository
          example: sample-repo
          required: true
          schema:
            type: string
        - name: include
          in: query
          description: |-
            Semicolon `;` separated list of qualified names of properties that need to be included in
            the response. If missing, the server selects a list of properties which have a low cost
            for the backing content management system as well as a high probability of being useful.
            
            The special string "default"  triggers the inclusion of those properties even if this
            parameter is given, making the properties named in  this list additiove. Otherwise, only
            the properties in this list are returned.
            
            Each property name can be prefixed with the next token of an object enumeration, if this
            property is a multi valued reference property. This token is ignored otherwise. The prefix
            is separated by the at `@` character.
            
            The semicolon `;` and at `@` characters within the qualified name or next token are
            escaped with the backslash `\` character.
          example: opendma:id;opendma:class;2GE3E4A@opendma:ContentElements;special:foo\;bar;default
          required: false
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Object'
  /opendma/obj/{repoId}/{objId}:
    get:
      tags:
        - Objects
      summary: Retrieve Object
      description: Returns an object.
      operationId: getObject
      parameters:
        - name: repoId
          in: path
          description: ID of Repository
          example: sample-repo
          required: true
          schema:
            type: string
        - name: objId
          in: path
          description: ID of Object
          example: 0ab4ca42-a0d9-4cfa-a2a3-30516c78203e
          required: true
          schema:
            type: string
        - name: include
          in: query
          description: |-
            Semicolon `;` separated list of qualified names of properties that need to be included in
            the response. If missing, the server selects a list of properties which have a low cost
            for the backing content management system as well as a high probability of being useful.
            
            The special string "default"  triggers the inclusion of those properties even if this
            parameter is given, making the properties named in  this list additiove. Otherwise, only
            the properties in this list are returned.
            
            Each property name can be prefixed with the next token of an object enumeration, if this
            property is a multi valued reference property. This token is ignored otherwise. The prefix
            is separated by the at `@` character.
            
            The semicolon `;` and at `@` characters within the qualified name or next token are
            escaped with the backslash `\` character.
          example: opendma:id;opendma:class;2GE3E4A@opendma:ContentElements;special:foo\;bar;default
          required: false
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Object'
  /opendma/bin/{repoId}/{contentId}:
    get:
      tags:
        - Content
      summary: Read binary data stream
      description: Returns an object.
      operationId: getContent
      parameters:
        - name: repoId
          in: path
          description: ID of Repository
          example: sample-repo
          required: true
          schema:
            type: string
        - name: contentId
          in: path
          description: ID of Content Stream
          example: 0ab4ca42-a0d9-4cfa-a2a3-30516c78203e
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
components:
  schemas:
    ServiceDescriptor:
      type: object
      description: Information about this service
      required:
        - opendmaVersion
        - serviceVersion
        - repositories
      properties:
        opendmaVersion:
          type: string
          example: 1.0.0
          description: Version of OpenDMA spec supported by this service
        serviceVersion:
          type: string
          example: 1.0.0
          description: Version of this Service
        repositories:
          type: array
          description: List of Repositories provided by this service
          items:
            type: string
            example: sample-repo
            description: ID of repository provided by this service
    Object:
      type: object
      description: |-
        Representation of a single OpenDMA object. Can be nested as the value of a property.
      properties:
        id:
          type: string
          example: 0ab4ca42-a0d9-4cfa-a2a3-30516c78203e
        rootOdmaClassName:
          type: string
          example: opendma:Object
        aspectRootOdmaNames:
          type: array
          items:
            type: string
            example: opendma:Document
        properties:
          type: array
          items:
            $ref: '#/components/schemas/Property'
        complete:
          type: boolean
    Property:
      type: object
      properties:
        name:
          type: string
          example: opendma:Title
        type:
          type: string
          example: String
          enum:
            - String
            - Integer
            - Short
            - Long
            - Float
            - Double
            - Boolean
            - DateTime
            - Blob
            - Reference
            - Content
            - Id
            - Guid
        multiValue:
          type: boolean
        readOnly:
          type: boolean
        resolved:
          type: boolean
        value:
          oneOf:
            - type: string
            - $ref: "#/components/schemas/ContentValue"
            - $ref: "#/components/schemas/GuidValue"
            - $ref: "#/components/schemas/Object"
            - type: array
              items:
                type: string
            - type: array
              items:
                $ref: "#/components/schemas/ContentValue"
            - type: array
              items:
                $ref: "#/components/schemas/GuidValue"
            - type: array
              items:
                $ref: "#/components/schemas/Object"
    ContentValue:
      type: object
      properties:
        id:
          type: string
        size:
          type: integer
          format: int64
    GuidValue:
      type: object
      properties:
        objectId:
          type: string
          example: 0ab4ca42-a0d9-4cfa-a2a3-30516c78203e
        repositoryId:
          type: string
          example: sample-repo